[MODIFIED] Review Feedback
1- If a book is on a bookshelf (Currently Reading, Want to Read, or Read), that should be reflected in both locations (search page and home page).

If a book is assigned to any shelf on the main page and that book also appears on the search page, then the correct shelf for any such book should be already selected on the search page. For all other books, the option none shall be selected.

2- Emptying the search input should remove the books which are being shown on the search page.
Currently, the books from the last search are being shown on the search page even when the search input box is empty.

3- Users are able to search books with multiple words as well.